requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
gunicorn
django_heroku
django-livereload-server==0.3.2
pycodestyle
graphene_django
graphene
django-graphql-auth
django-graphql-jwt